feat(auth): extract settings utility functions
