+++ feature/auth/src/main/kotlin/timur/gilfanov/messenger/auth/ui/screen/login/LoginScreen.kt
@@ -3,9 +3,6 @@
 package timur.gilfanov.messenger.auth.ui.screen.login
 
 import android.content.Context
-import android.content.Intent
-import android.net.Uri
-import android.provider.Settings
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
@@ -60,6 +57,8 @@ import timur.gilfanov.messenger.auth.R
 import timur.gilfanov.messenger.auth.ui.BlockingErrorDialog
 import timur.gilfanov.messenger.auth.ui.GoogleSignInClient
 import timur.gilfanov.messenger.auth.ui.GoogleSignInResult
+import timur.gilfanov.messenger.auth.ui.utils.openAppSettings
+import timur.gilfanov.messenger.auth.ui.utils.openStorageSettings
 import timur.gilfanov.messenger.auth.ui.utils.tooManyAttemptsDisplayString
 import timur.gilfanov.messenger.auth.validation.CredentialsValidatorImpl
 import timur.gilfanov.messenger.domain.entity.auth.validation.CredentialsValidationError
@@ -159,7 +158,7 @@ private fun LoginEffectHandler(
                         val message = effect.message
                         val actionLabel =
                             if (message is LoginSnackbarMessage.StorageTemporarilyUnavailable) {
-                                context.getString(R.string.login_action_retry)
+                                context.getString(R.string.snackbar_action_retry)
                             } else {
                                 null
                             }
@@ -397,36 +396,21 @@ private fun LoginGeneralError.toDisplayString(): String = when (this) {
 
 private fun LoginSnackbarMessage.toDisplayString(context: Context): String = when (this) {
     LoginSnackbarMessage.NetworkUnavailable ->
-        context.getString(R.string.login_error_network_unavailable)
+        context.getString(R.string.snackbar_error_network_unavailable)
 
     LoginSnackbarMessage.ServiceUnavailable ->
-        context.getString(R.string.login_error_service_unavailable)
+        context.getString(R.string.snackbar_error_service_unavailable)
 
     LoginSnackbarMessage.Unknown ->
-        context.getString(R.string.login_error_unknown)
+        context.getString(R.string.snackbar_error_unknown)
 
     LoginSnackbarMessage.GoogleSignInFailed ->
-        context.getString(R.string.login_error_google_sign_in_failed)
+        context.getString(R.string.snackbar_error_google_sign_in_failed)
 
     LoginSnackbarMessage.StorageTemporarilyUnavailable ->
-        context.getString(R.string.login_error_storage_temporarily_unavailable)
+        context.getString(R.string.snackbar_error_storage_temporarily_unavailable)
 
     is LoginSnackbarMessage.TooManyAttempts -> tooManyAttemptsDisplayString(context, remaining)
-}
-
-private fun openAppSettings(context: Context) {
-    val intent = Intent(Settings.ACTION_APPLICATION_DETAILS_SETTINGS).apply {
-        data = Uri.fromParts("package", context.packageName, null)
-    }
-    context.startActivity(intent)
-}
-
-private fun openStorageSettings(context: Context) {
-    val intent = Intent(Settings.ACTION_INTERNAL_STORAGE_SETTINGS)
-    if (intent.resolveActivity(context.packageManager) == null) {
-        intent.action = Settings.ACTION_SETTINGS
-    }
-    context.startActivity(intent)
 }
 
 @Preview
